+++ src/main/java/emi/mtg/deckbuilder/view/search/omnifilter/filters/RulesText.java
@@ -3,9 +3,9 @@ package emi.mtg.deckbuilder.view.search.omnifilter.filters;
 import emi.mtg.deckbuilder.model.CardInstance;
 import emi.mtg.deckbuilder.view.search.omnifilter.Omnifilter;
 
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.function.Predicate;
+import java.util.*;
+import java.util.function.*;
+import java.util.stream.Collectors;
 
 public class RulesText implements Omnifilter.Subfilter {
 	@Override
@@ -18,15 +18,25 @@ public class RulesText implements Omnifilter.Subfilter {
 		return "Search cards' rules (oracle) text. Only responds to the `:` operator.";
 	}
 
+	private static BiPredicate<String, String> create(String searchText) {
+		if (searchText.indexOf('~') < 0 && !searchText.contains("CARDNAME")) {
+			return (rules, name) -> rules.toLowerCase().contains(searchText.toLowerCase());
+		}
+
+		List<MatchUtils.SequenceMatcher> rope = Arrays.stream(searchText.split("(~|CARDNAME)"))
+				.filter(s -> !s.isEmpty())
+				.map(String::toLowerCase)
+				.map(MatchUtils::charSequenceMatcher)
+				.collect(Collectors.toList());
+
+		return (rules, name) -> MatchUtils.matchesRope(rules, 0, new MatchUtils.Delimited<>(rope, MatchUtils.cardNameOrSubNameMatcher(name))) >= 0;
+	}
+
 	@Override
 	public Predicate<CardInstance> create(Omnifilter.Operator operator, String value) {
 		if (operator != Omnifilter.Operator.DIRECT) throw new IllegalArgumentException("Can only use ':' with rules filter.");
 
-		if (value.indexOf('~') >= 0 || value.contains("CARDNAME")) {
-			return (Omnifilter.FaceFilter) face -> face.rules().toLowerCase().contains(value.replace("~", face.name()).replace("CARDNAME", face.name()).toLowerCase());
-		} else {
-			final String search = value.toLowerCase();
-			return (Omnifilter.FaceFilter) face -> face.rules().toLowerCase().contains(search);
-		}
+		BiPredicate<String, String> base = RulesText.create(value);
+		return (Omnifilter.FaceFilter) face -> base.test(face.rules(), face.name());
 	}
 }